Determine resource usages between project library dependencies.
The dependency analyzer will now account for resource references between
libraries, i.e a resource declared in library A is referenced in libraryB.
It determines if a resource is used by building using resource declarations and reference
symbols from each library and then mapping the symbol to the dependency, creating a graph.
This does not take application modules into account - this will be worked on in
following changes.

Test: AnalyzeDependenciesTaskUtilsTest
Bug: 77246430
Change-Id: Idb1b351e462831a8105dd436b1701a7209fea80c