AppId / PackageName via new Manifest parser.
This change makes the applicationId be via
Provider throughout the codebase, and change
how we handle the tested applicationId as well
as it was quite a mess before.

This also renames originalAppId to packageName
as this is really what this is.

This also removed usage of the old manifest parser.
The classes are still there but they are now
@Deprecated

Bug: N/A (refactoring)
Test: existing tests
Change-Id: Ide8e7e715f36eb0fa65e6a485c09c318bbfcfd24