Erase timestamps of zip entries when merging Java resources
For correct UP-TO-DATE and cacheability checks, usually either the
producer of a zip file needs to erase the zip entries' timestamps, or
the consumer of the zip file needs to ignore the zip entries'
timestamps, so that two zip files with the same contents but with
different zip entries' timestamps are still considered the same by
Gradle.

In the bug below, the consumer is not yet able to do its part, so the
producer (MergeJavaResourcesTask) will need to do its part.

Bug: 142890134
Test: Updated unit test MergeJavaResourcesDelegateTest
      More integration tests are tracked at bug 143357824
Change-Id: I1cf651e2e3e7993038b531d0360afeef943ca455