Call Android Test Retention plugin in AGP
We had a previous discussion to enable Android Test Retention (previous
code name Icebox) in AGP, then we got the suggestion that we should move
the Retention implementation into Unified Test Platform (UTP) and wire
it back to AGP through UTP. This CL does exactly that.

Currently Retention was implemented as a plugin in UTP. This CL
introduces an experimental flag to enable the Retention plugin. The
external flag is:

android.experimental.androidTest.useUnifiedTestPlatform.useRetention

Bug: 152993451
Test: UtpConfigFactoryTest
Change-Id: I9193c4c2ed767c6194fcf111e57785c18c1857ed