TaskAction -> TaskConfigurationActions
The name was confusing because it has nothing to do with
the actual task actions. It's about wrapping all the configuration
action into a single Gradle Action<> to pass to the task container

This is particularly useful now that we have chrome tracing of
sync available in Android and users can see the calls to this
class.

Bug: N/A
Test: existing tests.
Change-Id: I07742a5814e05d51cf6dcf8f24463ebe9d57d93f