Add integration test for ProcessApplicationManifest
A bug introduced an issue, in another branch, wherein
changes to versionCode in build.gradle were not always
reflected in the manifest included in the user's apk
or bundle. This change cherry-picks the regression
test added for that bug, even though the issue did not
need to be fixed on this branch.

Bug: b/148525428
Test: Added regression test in ManifestMergingTest
Change-Id: I38c69a765367e5dc119a4e394722668957322291
(cherry picked from commit 85ad34a16b55e6c99b2f04ff93380153558bdad9)