Change AGP to tolerate in-house built CMake version number
At some point, CMake build changed to produce a special version
number when built externally from Kitware:

  3.17.0-gc5272a5

This kind of version number isn't supported by the Revision class
we use to parse version numbers.

This change is to preprocess the string to remove the hash code
from the version before passing it to Revision.parseRevision.

One consequence of this issue is that prior versions of AGP
will not be able to tolerate the CMake 3.17.0 in our SDK. That's
pretty unfortunate.

Bug: 152223150
Test: Added unittests
Change-Id: I79d2884740ef6e10eb51b40c8c78161a876cc8b2