Avoid using .map on providers that are task outputs
This is causing failures with --warning-mode=fail in
Gradle 6.4-rc-2. AGP must not invoke .map on providers
that are task outputs before those tasks run.

This CL removes all instances of
getFinalProductAsFileCollection as that method was
using .map. Instead, all getFinalProduct invocation
are guarded using the same logic as for the creation
of producers.

Additionally, JSON configuration for core library desugaring
is not using a custom ValueSource. This is because previous
implementation was causing an issue similar to b/154315491.

There is an issue with using finalizeValueOnRead() tracked in
https://github.com/gradle/gradle/issues/12811 so this CL
removes this when creating AGP providers.

Bug: 154315491, updating to Gradle 6.4-rc-2
Test: existing
Change-Id: I9ab652992a4b44b15d6f4d36f547a48da196518b