Use Gradle Project callbacks to enforce the same AGP version
This CL uses Gradle.allprojects in order to enforce the same
AGP version across projects. This mechanism avoids the need
for MBean that was used to share this information.

Current approach will register the check for the first project
that invokes it in the main build. As projects are evaluated
we will get notified if AGP is applied. This is implemented
for composite buids as well. As soon as violation is found
the issue is reported.

Although this CL relies on casting IncludedBuild to
IncludedBuildState (an internal Gradle type), this is
already used in tooling API (IdeaModelBuilder) and we should
not expect it to change. Otherwise, only public APIs are used.

Unused classes used to support previous version checking
mechanims are removed in this CL.

Bug: 154742527
Test: CompositeBuildTest, EnforceTheSameAgpVersionsTest
Change-Id: If149becea3a2d5bc301cccebc8f1713cd0d0a58f