Enable dynamic scheduling on Windows machines.
We currently use dynamic scheduling on linux and have seen a 40-60%
improvement in javac and kotlinc build times. Dynamic scheduling allows
the host machine to run the same actions locally while they are also
sent to RBE. Javac and kotlinc actions can make use of persistent
workers locally, reducing the overhead of initializing a JVM and
compilation setup.

The fallback local strategies (sandboxed, local) are removed since we
don't need them. And because 'sandboxed' causes bazel to fail on Windows
as true sandboxing is not supported on Windows.

Bug: 147615404
Test: Reduced javac/kotlinc action duration in the bazel profile.
Change-Id: I9b7f6e7350b1175f609decffc44ae1c44366af0b