Legacy variant API: Make MergedFlavor maps mutable
As the legacy variant API uses the Gradle Tooling API interfaces,
it gets generally non-mutable view of the world. However a few things
can be changed on the merged flavor, including two map-like things,
manifestPlaceholders and testInstrumentationRunnerArguments.

This change introduces a new intermediate interface to expose those in kotlin as MutableMaps.

Fixes: https://b.corp.google.com/issues/155318103
Test: Expanded KotlinDslTest, updated pseudo-api expectation file.
Change-Id: Ib9dd9f4598611c052405fa5ce9be1edcf35537d5