Generate individual baseline XML and LCOV
Compute baseline coverage for each source target by feeding the class
jar of source files to Jacoco to produce an XML report and then convert
it to LCOV using the target's filtered baseline source file list.

Since the jars involved are small and already built this executes
quickly. It also minimizes the amount of work we need to run through
Jacoco to just the incremental changes.

Change-Id: I9ff06500863f614e4ac6b900fe124d14d9097f95
Test: N/A
Bug: 153913236